localizes styles
